Move preformatted block scanner to format-independent location
