+++ crates/ty_ide/src/docstring/rest.rs
@@ -5,7 +5,7 @@ use ruff_python_trivia::leading_indentation;
 use ruff_source_file::{Line as SourceLine, UniversalNewlineIterator, UniversalNewlines};
 use ruff_text_size::{TextRange, TextSize};
 
-use super::markdown;
+use super::preformatted::PreformattedBlockScanner;
 
 /// Represents a parsed restructured text (reST) docstring.
 pub(super) struct Docstring {
@@ -120,7 +120,7 @@ impl FieldList {
             }
 
             let Some(field_list) = Self::parse(&mut lines) else {
-                preformatted_blocks.observe_non_field_line(line.text);
+                preformatted_blocks.observe_non_preformatted_line(line.text);
                 lines.next();
                 continue;
             };
@@ -237,225 +237,6 @@ impl FieldList {
         FieldHeader::indentation(non_blank_line.text) > indent
             || FieldHeader::at_indent(non_blank_line.text, indent).is_some()
     }
-}
-
-/// Recognizes preformatted blocks that may occur within a docstring (e.g. a markdown fence).
-#[derive(Debug, Default, Clone, PartialEq, Eq)]
-struct PreformattedBlockScanner<'a> {
-    active_markdown_fence: Option<markdown::MarkdownFence<'a>>,
-    active_doctest: bool,
-    preformatted_block_state: PreformattedBlockState,
-}
-
-/// The set of characters that can each be used to denote a block quote.
-///
-/// <https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#quoted-literal-blocks>
-const QUOTED_LITERAL_BLOCK_QUOTE_CHARACTERS: &str = r##"!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~"##;
-
-impl<'a> PreformattedBlockScanner<'a> {
-    /// Updates internal state to reflect the given line and returns whether or
-    /// not the given line is contained within a preformatted block.
-    fn consume_preformatted_line(&mut self, line: &'a str) -> bool {
-        if let Some(fence) = self.active_markdown_fence {
-            if fence.is_closed_by(line) {
-                self.active_markdown_fence = None;
-            }
-            return true;
-        }
-
-        if self.is_within_preformatted_block(line) {
-            return true;
-        }
-
-        let trimmed = line.trim_start_matches(' ');
-        if self.active_doctest {
-            if trimmed.is_empty() {
-                self.active_doctest = false;
-            }
-            return true;
-        }
-
-        if trimmed.starts_with(">>>") {
-            self.active_doctest = true;
-            return true;
-        }
-
-        if let Some(fence) = markdown::MarkdownFence::find(line) {
-            self.active_markdown_fence = Some(fence);
-            return true;
-        }
-
-        false
-    }
-
-    /// Whether or not the given line is specifically within a preformatted block
-    /// introduced by reST syntax.
-    fn is_within_preformatted_block(&mut self, line: &str) -> bool {
-        let current_indent = FieldHeader::indentation(line);
-        let line_is_empty = line.trim_start().is_empty();
-
-        match self.preformatted_block_state {
-            PreformattedBlockState::Active(PreformattedBlockKind::Indented { marker_indent }) => {
-                if !line_is_empty && current_indent <= marker_indent {
-                    // We've reached the de-dent that marks the end of the preformatted block.
-                    self.preformatted_block_state = PreformattedBlockState::Inactive;
-                    false
-                } else {
-                    true
-                }
-            }
-            PreformattedBlockState::Active(PreformattedBlockKind::QuotedLiteral {
-                indent,
-                quote,
-            }) => {
-                if line_is_empty {
-                    self.preformatted_block_state = PreformattedBlockState::Inactive;
-                    false
-                } else if Self::quote_character(line, indent) == Some(quote) {
-                    true
-                } else {
-                    self.preformatted_block_state = PreformattedBlockState::Inactive;
-                    false
-                }
-            }
-            PreformattedBlockState::Pending {
-                marker_indent,
-                allows_quoted_literal_block,
-            } if !line_is_empty => {
-                if current_indent > marker_indent {
-                    // We just entered a new preformatted block.
-                    self.preformatted_block_state =
-                        PreformattedBlockState::Active(PreformattedBlockKind::Indented {
-                            marker_indent,
-                        });
-                    true
-                } else if allows_quoted_literal_block
-                    && let Some(quote) = Self::quote_character(line, marker_indent)
-                {
-                    self.preformatted_block_state =
-                        PreformattedBlockState::Active(PreformattedBlockKind::QuotedLiteral {
-                            indent: marker_indent,
-                            quote,
-                        });
-                    true
-                } else {
-                    self.preformatted_block_state = PreformattedBlockState::Inactive;
-                    false
-                }
-            }
-            PreformattedBlockState::Pending { .. } | PreformattedBlockState::Inactive => false,
-        }
-    }
-
-    /// Updates internal state that allows us to detect preformatted blocks introduced by reST
-    /// syntax.
-    fn observe_non_field_line(&mut self, line: &str) {
-        if matches!(
-            self.preformatted_block_state,
-            PreformattedBlockState::Inactive
-        ) && Self::starts_preformatted_block(line.trim_start())
-        {
-            self.preformatted_block_state = PreformattedBlockState::Pending {
-                marker_indent: FieldHeader::indentation(line),
-                allows_quoted_literal_block: Self::allows_quoted_literal_block(line.trim_start()),
-            };
-        }
-    }
-
-    /// Whether or not the given line marks the start of a preformatted block.
-    fn starts_preformatted_block(line: &str) -> bool {
-        let Some(marker) = Self::preformatted_block_marker(line) else {
-            return false;
-        };
-
-        !matches!(
-            marker,
-            PreformattedBlockMarker::Directive(
-                "attention"
-                    | "caution"
-                    | "danger"
-                    | "error"
-                    | "hint"
-                    | "important"
-                    | "note"
-                    | "tip"
-                    | "warning"
-                    | "admonition"
-                    | "seealso"
-                    | "versionadded"
-                    | "version-added"
-                    | "versionchanged"
-                    | "version-changed"
-                    | "version-deprecated"
-                    | "deprecated"
-                    | "version-removed"
-                    | "versionremoved"
-            )
-        )
-    }
-
-    /// Tries to identify a marker that introduces a preformatted block.
-    fn preformatted_block_marker(line: &str) -> Option<PreformattedBlockMarker<'_>> {
-        let marker = if let Some(marker) = line.strip_suffix("::") {
-            marker
-        } else {
-            let (before_language, _language) = line.rsplit_once(' ')?;
-            before_language.trim_end().strip_suffix("::")?
-        };
-
-        if let Some(directive) = marker.strip_prefix(".. ") {
-            Some(PreformattedBlockMarker::Directive(directive))
-        } else {
-            Some(PreformattedBlockMarker::Paragraph)
-        }
-    }
-
-    /// Whether or not a particular preformatted block can contain an unindented quoted literal block.
-    fn allows_quoted_literal_block(line: &str) -> bool {
-        line.ends_with("::")
-            && matches!(
-                Self::preformatted_block_marker(line),
-                Some(PreformattedBlockMarker::Paragraph)
-            )
-    }
-
-    /// Returns the quote character for a quoted literal block line.
-    fn quote_character(line: &str, indent: TextSize) -> Option<char> {
-        if FieldHeader::indentation(line) != indent {
-            return None;
-        }
-
-        let quote = line.get(indent.to_usize()..)?.chars().next()?;
-        QUOTED_LITERAL_BLOCK_QUOTE_CHARACTERS
-            .contains(quote)
-            .then_some(quote)
-    }
-}
-
-/// Identifies the syntax that introduced a potential preformatted block.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum PreformattedBlockMarker<'a> {
-    Paragraph,
-    Directive(&'a str),
-}
-
-/// Tracks the state of a preformatted block introduced by reST syntax.
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
-enum PreformattedBlockState {
-    #[default]
-    Inactive,
-    Pending {
-        marker_indent: TextSize,
-        allows_quoted_literal_block: bool,
-    },
-    Active(PreformattedBlockKind),
-}
-
-/// Tracks the type of an active preformatted block.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum PreformattedBlockKind {
-    Indented { marker_indent: TextSize },
-    QuotedLiteral { indent: TextSize, quote: char },
 }
 
 /// Constructs new instances of the model for a reST field.
